MonsterFirstとMonsterIgnoreを追加

+++ src/char/char.cpp
@@ -1735,9 +1735,11 @@ enum e_char_del_response char_delete(struct char_session_data* sd, uint32 char_i
 	if( SQL_ERROR == Sql_Query(sql_handle, "DELETE FROM `pybot_cart_auto_get_item` WHERE `char_id`='%d'", char_id) ||
 		SQL_ERROR == Sql_Query(sql_handle, "DELETE FROM `pybot_distance_policy` WHERE `char_id`='%d'", char_id) ||
 		SQL_ERROR == Sql_Query(sql_handle, "DELETE FROM `pybot_equipset` WHERE `char_id`='%d'", char_id) ||
+		SQL_ERROR == Sql_Query(sql_handle, "DELETE FROM `pybot_first_mob` WHERE `char_id`='%d'", char_id) ||
 		SQL_ERROR == Sql_Query(sql_handle, "DELETE FROM `pybot_first_skill` WHERE `char_id`='%d'", char_id) ||
 		SQL_ERROR == Sql_Query(sql_handle, "DELETE FROM `pybot_great_mob` WHERE `char_id`='%d'", char_id) ||
 		SQL_ERROR == Sql_Query(sql_handle, "DELETE FROM `pybot_ignore_item` WHERE `char_id`='%d'", char_id) ||
+		SQL_ERROR == Sql_Query(sql_handle, "DELETE FROM `pybot_ignore_mob` WHERE `char_id`='%d'", char_id) ||
 		SQL_ERROR == Sql_Query(sql_handle, "DELETE FROM `pybot_limit_skill` WHERE `char_id`='%d'", char_id) ||
 		SQL_ERROR == Sql_Query(sql_handle, "DELETE FROM `pybot_memo` WHERE `char_id`='%d'", char_id) ||
 		SQL_ERROR == Sql_Query(sql_handle, "DELETE FROM `pybot_mvp` WHERE `char_id`='%d'", char_id) ||
